checked for missing pid, adjusted cmatrix settings

+++ ramCheck.sh
@@ -1,11 +1,10 @@
 #!/bin/bash
 
-#sh -c 'L=$$; echo $L;exec cmatrix -b -s&'
+#sh -c 'L=$$; echo $L;exec cmatrix -b -s&' #initial test to get pid
 
 while true; do
-	#mt=$(free -m | awk 'NR==2{printf "Mem: %s/%sMB (%.2f%%)\n", $2-($4),$2,($2-($4))*100/$2 }')
+	#mt=$(free -m | awk 'NR==2{printf "Mem: %s/%sMB (%.2f%%)\n", $2-($4),$2,($2-($4))*100/$2 }') #getting memory
 	mt=$(free -m | awk 'NR==2{printf "%i", ($2-($4))*100/$2 }')
-	#prev=""
 	case 1 in
 		$(($mt > 95)))
 			c="red"
@@ -23,14 +22,17 @@ while true; do
 			c="white"
 			;;
 	esac
-	#echo "$mt | $c | $prev"
+	#echo "$mt | $c | $prev" #debug
 	sleep 5
 	if [[ $prev != $c ]] 
 	then
 		prev=$c
-		#echo "changed!"
-		kill -15 $bg_PID
-		cmd="cmatrix -C $c"
+		#echo "changed!" #debug
+		if [[ $bg_PID != "" ]] 
+		then
+			kill -15 $bg_PID
+		fi
+		cmd="cmatrix -n -a -C $c"
 		$cmd &
 		bg_PID=$!
 	fi
